Finished inanimate class
Inanimate class finished. Can be used, but probably not for anything
useful. For example, you CAN use this object for platforms, but it would
be a bit more complicated to control its effect on the character.
Instead, I am making platforms its own class which will be able to stop
the character from falling through it without any logic in the main
program itself. It will be handled within the class itself.