Created animated sprites class
This class will allow you to create objects with ANIMATED sprites
(unlike the inanimate class, which only has one picture as the sprite).
Please read the class AND it's comments (especially the constructor!!!)
so you know what kind of sprites you need to give it in order to create
a working sprite.