Minor improvements and fixes to classes + added Enemy object
You may subclass the abstract class Enemy and add your own properties
and methods to it to make the enemies specific. Also fixed some
important stuff. Changed velocity and gravity and stuff in the animate
class and its subclasses to double so that you can have more fine
control over the speed (so you don't have to just deal with big numbers)