Fixed animate out of index bug of sprites in animate class, minor changes
In player class, two new boolean properties have been added to indicate
that the player is currently moving left or right.
In the animate class, the intState property has been made private. Refer
to javadoc to learn how to use the new methods that will let you use
this property instead of using it directly